update: separate is_allocated.c
