+++ srcs/free.c
@@ -9,38 +9,6 @@
 #include "utils.h"
 #include <sys/mman.h>
 
-static bool	_is_allocated_block(t_malloc_chunk *region, t_malloc_chunk *address)
-{
-	t_malloc_chunk	*chunk;
-
-	chunk = region;
-	while (chunk >= address && !IS_FOOTER(chunk))
-	{
-		if (chunk == address)
-			return (IS_ALLOCED(chunk));
-		chunk = NEXTCHUNK(chunk);
-	}
-	return (false);
-}
-
-static bool	_is_allocated(t_magazine *magazine, t_malloc_chunk *address)
-{
-	t_malloc_chunk	*footer;
-	
-	if (magazine->regions <= address && address < magazine->top)
-		return (_is_allocated_block(magazine->regions, address));
-
-	footer = (void *)magazine->regions + magazine->size - REGION_FOOTERSIZE;
-	for (t_malloc_chunk *region = footer->fd; region; )
-	{
-		footer = (void *)region + magazine->size - REGION_FOOTERSIZE;
-		if (region <= address && address < footer)
-			return (_is_allocated_block(region, address));
-		region = footer->fd;
-	}
-	return (false);
-}
-
 static void	_free_alloc(t_magazine *magazine, t_malloc_chunk *chunk)
 {
 	t_malloc_chunk	*next;
